# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Command API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  systemHealth: number;
  activeProjects: number;
  pendingTasks: number;
  pendingApprovals: number;
  securityScore: number;
  activeAgents: number;
  unreadMessages: number;
  todayMeetings: number;
  currentMode: string;
}

export interface ChartDataPoint {
  label: string;
  value: number;
}

export interface AgentPerformancePoint {
  name: string;
  reliability: number;
  tasks: number;
  cost: number;
}

export interface PieDataPoint {
  name: string;
  value: number;
  color: string;
}

export interface ProjectProgressPoint {
  name: string;
  progress: number;
  status: string;
}

export interface AnalyticsData {
  taskCompletion: ChartDataPoint[];
  agentPerformance: AgentPerformancePoint[];
  communicationVolume: ChartDataPoint[];
  securityTrends: ChartDataPoint[];
  timeAllocation: PieDataPoint[];
  projectProgress: ProjectProgressPoint[];
}

export interface Agent {
  id: number;
  name: string;
  description: string;
  category: string;
  reliabilityScore: number;
  costScore: number;
  status: string;
  supportedModels: string[];
  strengths: string[];
  useCases: string[];
  lastUpdated?: string;
  permissions?: string[];
  tasksCompleted?: number;
  avgResponseTime?: number;
}

export interface TaskRouteInput {
  task: string;
}

export interface TaskRouteResult {
  primaryAgent: Agent;
  secondaryAgent: Agent;
  estimatedCost: number;
  confidenceScore: number;
  reasoning: string;
}

export interface InboxItem {
  id: number;
  source: string;
  sourceIcon?: string;
  sender?: string;
  subject: string;
  summary: string;
  priority: string;
  classification: string;
  aiRecommendation?: string;
  relatedProject?: string;
  priorityScore?: number;
  isRead: boolean;
  isArchived?: boolean;
  createdAt: string;
}

export interface InboxItemUpdate {
  isRead?: boolean;
  isArchived?: boolean;
  classification?: string;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  owner: string;
  status: string;
  priority: string;
  progress: number;
  risks?: string[];
  tags?: string[];
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  taskCount?: number;
  completedTaskCount?: number;
}

export interface ProjectInput {
  name: string;
  description: string;
  owner: string;
  status: string;
  priority: string;
  /** @nullable */
  dueDate?: string | null;
  tags?: string[];
}

export interface ProjectUpdate {
  name?: string;
  description?: string;
  status?: string;
  priority?: string;
  progress?: number;
  /** @nullable */
  dueDate?: string | null;
}

export interface Task {
  id: number;
  title: string;
  description?: string;
  status: string;
  priority: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  assignee?: string;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  aiRecommendation?: string;
}

export interface TaskInput {
  title: string;
  description?: string;
  status: string;
  priority: string;
  /** @nullable */
  projectId?: number | null;
  assignee?: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface TaskUpdate {
  title?: string;
  status?: string;
  priority?: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface CalendarEvent {
  id: number;
  title: string;
  description?: string;
  startTime: string;
  endTime: string;
  type: string;
  attendees?: string[];
  location?: string;
  aiPrep?: string;
  relatedProject?: string;
  actionItems?: string[];
}

export interface Communication {
  id: number;
  sender: string;
  senderEmail?: string;
  subject: string;
  source: string;
  summary: string;
  sentiment: string;
  priority: string;
  suggestedResponse?: string;
  relatedProject?: string;
  requiredActions?: string[];
  isRead?: boolean;
  createdAt: string;
}

export interface CommunicationAnalytics {
  totalMessages: number;
  responseRate: number;
  avgResponseTime: number;
  bySource: PieDataPoint[];
  bySentiment: PieDataPoint[];
}

export interface Automation {
  id: number;
  name: string;
  description: string;
  trigger: string;
  status: string;
  applicationsUsed?: string[];
  requiredPermissions?: string[];
  requiresApproval?: boolean;
  lastExecution: string;
  timeSaved?: number;
  runsCount?: number;
}

export interface AutomationRunResult {
  success: boolean;
  message: string;
  executedAt: string;
}

export interface SecurityLayer {
  id: number;
  name: string;
  description: string;
  status: string;
  score: number;
  lastScan?: string;
}

export interface SecurityStatus {
  overallScore: number;
  threatLevel: string;
  activeThreats?: number;
  vulnerabilities?: number;
  complianceStatus?: string;
  vaultStatus?: string;
  layers: SecurityLayer[];
}

export interface SecurityEvent {
  id: number;
  type: string;
  severity: string;
  description: string;
  timestamp: string;
  resolved?: boolean;
}

export interface VaultItem {
  id: number;
  name: string;
  category: string;
  permissionLevel: string;
  lastAccessed: string;
  owner: string;
  description?: string;
  status?: string;
}

export interface ResearchItem {
  id: number;
  title: string;
  description?: string;
  type: string;
  status: string;
  tags?: string[];
  sources?: number;
  credibilityScore?: number;
  aiSummary?: string;
  createdAt: string;
}

export interface MemoryItem {
  id: number;
  category: string;
  key: string;
  value: string;
  isProtected?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface MemoryItemInput {
  category: string;
  key: string;
  value: string;
  isProtected?: boolean;
}

export interface Approval {
  id: number;
  action: string;
  requestingAgent: string;
  reason: string;
  riskLevel: string;
  dataInvolved?: string;
  status: string;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

export interface Insight {
  id: number;
  message: string;
  category: string;
  confidenceScore: number;
  priority?: string;
  actionRequired?: boolean;
  relatedProject?: string;
  createdAt: string;
}

export interface OperatingMode {
  id: number;
  name: string;
  description: string;
  icon: string;
  isActive: boolean;
  color?: string;
}

export interface SetModeInput {
  modeId: number;
}

export interface MarketplaceAgent {
  id: number;
  name: string;
  description: string;
  provider: string;
  category: string;
  accuracyScore: number;
  costScore: number;
  speedScore: number;
  reliabilityScore: number;
  privacyScore: number;
  isInstalled: boolean;
  version?: string;
  pricing?: string;
}