# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus Command API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/route:
    post:
      operationId: routeTask
      tags: [agents]
      summary: Route a task to the best agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRouteInput"
      responses:
        "200":
          description: Routing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRouteResult"
...
  /inbox/{id}:
    patch:
      operationId: updateInboxItem
      tags: [inbox]
      summary: Update inbox item (mark read, archive, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboxItemUpdate"
      responses:
        "200":
          description: Updated inbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
...
  /automations/{id}/run:
    post:
      operationId: runAutomation
      tags: [automations]
      summary: Run an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRunResult"
...
  /memory:
    get:
      operationId: listMemoryItems
      tags: [memory]
      summary: List memory items
      responses:
        "200":
          description: Memory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryItem"
    post:
      operationId: createMemoryItem
      tags: [memory]
      summary: Create a memory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryItemInput"
      responses:
        "201":
          description: Created memory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryItem"
  /memory/{id}:
    delete:
      operationId: deleteMemoryItem
      tags: [memory]
      summary: Delete (forget) a memory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /approvals/{id}/approve:
    post:
      operationId: approveRequest
      tags: [approvals]
      summary: Approve a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [approvals]
      summary: Reject a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
...
  /modes/current:
    get:
      operationId: getCurrentMode
      tags: [modes]
      summary: Get current operating mode
      responses:
        "200":
          description: Current mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatingMode"
    put:
      operationId: setCurrentMode
      tags: [modes]
      summary: Set current operating mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetModeInput"
      responses:
        "200":
          description: Updated mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatingMode"
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, name, description, category, reliabilityScore, costScore, status, supportedModels, strengths, useCases]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        reliabilityScore:
          type: number
        costScore:
          type: number
        status:
          type: string
        supportedModels:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        useCases:
          type: array
          items:
            type: string
        lastUpdated:
          type: string
        permissions:
          type: array
          items:
            type: string
        tasksCompleted:
          type: integer
        avgResponseTime:
          type: number
    TaskRouteInput:
      type: object
      required: [task]
      properties:
        task:
          type: string
    TaskRouteResult:
      type: object
      required: [primaryAgent, secondaryAgent, estimatedCost, confidenceScore, reasoning]
      properties:
        primaryAgent:
          $ref: "#/components/schemas/Agent"
        secondaryAgent:
          $ref: "#/components/schemas/Agent"
        estimatedCost:
          type: number
        confidenceScore:
          type: number
        reasoning:
          type: string
    InboxItem:
      type: object
      required: [id, source, subject, summary, priority, classification, isRead, createdAt]
      properties:
        id:
          type: integer
        source:
          type: string
        sourceIcon:
          type: string
        sender:
          type: string
        subject:
          type: string
        summary:
          type: string
        priority:
          type: string
        classification:
          type: string
        aiRecommendation:
          type: string
        relatedProject:
          type: string
        priorityScore:
          type: integer
        isRead:
          type: boolean
        isArchived:
          type: boolean
        createdAt:
          type: string
    InboxItemUpdate:
      type: object
      properties:
        isRead:
          type: boolean
        isArchived:
          type: boolean
        classification:
          type: string
    Project:
      type: object
      required: [id, name, description, owner, status, priority, progress, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        owner:
          type: string
        status:
          type: string
        priority:
          type: string
        progress:
          type: integer
        risks:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
    ProjectInput:
      type: object
      required: [name, description, owner, status, priority]
      properties:
        name:
          type: string
        description:
          type: string
        owner:
          type: string
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        progress:
          type: integer
        dueDate:
          type: ["string", "null"]
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        assignee:
          type: string
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        aiRecommendation:
          type: string
...
    AutomationRunResult:
      type: object
      required: [success, message, executedAt]
      properties:
        success:
          type: boolean
        message:
          type: string
        executedAt:
          type: string
...
    MemoryItem:
      type: object
      required: [id, category, key, value, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
        key:
          type: string
        value:
          type: string
        isProtected:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    MemoryItemInput:
      type: object
      required: [category, key, value]
      properties:
        category:
          type: string
        key:
          type: string
        value:
          type: string
        isProtected:
          type: boolean
    Approval:
      type: object
      required: [id, action, requestingAgent, reason, riskLevel, status, createdAt]
      properties:
        id:
          type: integer
        action:
          type: string
        requestingAgent:
          type: string
        reason:
          type: string
        riskLevel:
          type: string
        dataInvolved:
          type: string
        status:
          type: string
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
...
    OperatingMode:
      type: object
      required: [id, name, description, icon, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
        color:
          type: string
    SetModeInput:
      type: object
      required: [modeId]
      properties:
        modeId:
          type: integer
...